backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1.0
python-multipart>=0.0.6
psutil>=5.9.0
numpy>=1.24.0
h5py>=3.7.0
Pillow>=10.0.0
fabio>=2023.3.0
scipy>=1.10.0
matplotlib>=3.5.0
pyFAI>=2023.6.0
